Simplify tunnel lifecycle internals
Narrow the container runtime-run control surface and centralize Durable
Object tunnel state mutation so lifecycle recovery paths stay exact and
easier to reason about.
